refactor: Datatable fix type imports with index
