+++ src/components/DataTable/Columns/index.ts
@@ -1,0 +1,2 @@
+export * from './dataTableColumns';
+export * from './dataTableColumns.types';
