+++ src/components/DataTable/CustomCells/index.ts
@@ -1,0 +1,2 @@
+export * from './useDataTableCustomCells';
+export * from './dataTableCustomCells.types';
