+++ src/components/DataTable/Pagination/dataTablePagination.ts
@@ -1,5 +1,5 @@
 import type {PaginationProps as ComponentPaginationProps} from '~/components/Pagination';
-import type {DataTablePaginationLogicProps} from './dataTablePagination.types';
+import type {DataTablePaginationLogicProps} from './index';
 
 export const getDataTablePaginationProps = <T extends NonNullable<unknown>>({
     table,
