+++ src/components/DataTable/Pagination/index.ts
@@ -1,0 +1,2 @@
+export * from './dataTablePagination';
+export * from './dataTablePagination.types';
