+++ src/components/DataTable/cells/TableCell/index.ts
@@ -1,0 +1,2 @@
+export * from './TableCell';
+export * from './TableCell.types';
