+++ src/components/DataTable/cells/TableCellActions/index.ts
@@ -1,0 +1,2 @@
+export * from './TableCellActions';
+export * from './TableCellActions.types';
