+++ src/components/DataTable/cells/TableCellStatus/index.ts
@@ -1,0 +1,2 @@
+export * from './TableCellStatus';
+export * from './TableCellStatus.types';
