+++ src/components/DataTable/cells/TableHeadCell/index.ts
@@ -1,0 +1,2 @@
+export * from './TableHeadCell';
+export * from './TableHeadCell.types';
