+++ src/components/DataTable/cells/TableStructuredCell/TableStructuredCell.types.ts
@@ -1,4 +1,4 @@
-import type {TableCellProps} from '../TableCell/TableCell.types';
+import type {TableCellProps} from '../TableCell';
 
 export type TableStructuredCellProps = TableCellProps & {
     /**
