+++ src/components/DataTable/cells/TableStructuredCell/index.ts
@@ -1,0 +1,2 @@
+export * from './TableStructuredCell';
+export * from './TableStructuredCell.types';
