+++ src/components/DataTable/cells/index.ts
@@ -1,11 +1,6 @@
-export * from './TableCell/TableCell';
-export * from './TableCell/TableCell.types';
-export * from './TableCellActions/TableCellActions';
-export * from './TableCellActions/TableCellActions.types';
-export * from './TableCellStatus/TableCellStatus';
-export * from './TableCellStatus/TableCellStatus.types';
-export * from './TableStructuredCell/TableStructuredCell';
-export * from './TableStructuredCell/TableStructuredCell.types';
-export * from './TableHeadCell/TableHeadCell';
-export * from './TableHeadCell/TableHeadCell.types';
+export * from './TableCell';
+export * from './TableCellActions';
+export * from './TableCellStatus';
+export * from './TableStructuredCell';
+export * from './TableHeadCell';
